Updates made for iDaaS-connect-ThirdParty code consolidation
