+++ iDaaS-Connect-ThirdParty/src/main/java/io/connectedhealth_idaas/parsers/ResearchData.java
@@ -1,4 +1,4 @@
-package io.connectedhealth_idaas.thirdparty;
+package com.redhat.idaas.connect.parsers;
 
 import org.apache.camel.dataformat.bindy.annotation.CsvRecord;
 import org.apache.camel.dataformat.bindy.annotation.DataField;
