+++ iDaaS-Connect-ThirdParty/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 # Admin Interface Settings
-management.endpoints.web.exposure.include=hawtio, jolokia
+management.endpoints.web.exposure.include=hawtio, jolokia,info, health, prometheus
 hawtio.authenticationEnabled=false
 management.endpoint.hawtio.enabled=true
 management.endpoint.jolokia.enabled=true
@@ -10,18 +10,16 @@ management.endpoint.jolokia.enabled=true
 # Must be unique and defined otherwise defaults to 8080
 # used for any Fuse SpringBoot developed assets
 server.port=9983
-# Kafka
-kafkaBrokers=localhost:9092
+idaas.kafkaBrokers=localhost:9092
+idaas.integrationTopic=kic_dataintgrtntransactions
+idaas.appintegrationTopic=kic_appintgrtntransactions
+idaas.iotintegrationTopic=iot
 # JDBC Database
-#spring.datasource.url=jdbc:mysql://localhost/idaas
-spring.datasource.url=jdbc:mysql://localhost/idaas?useLegacyDatetimeCode=false&serverTimezone=America/Chicago
-#jdbc:postgresql://localhost:5432/idaas
-spring.datasource.username=idaas
-spring.datasource.password=@idaas123
-#spring.database.driver-class-name=com.mysql.cj.jdbc.Driver
-#org.postgresql.Driver
+spring.datasource.url=jdbc:postgresql://localhost/:5432/idaas_thirdparty
+spring.datasource.username=postgres
+spring.datasource.password=Developer123
+spring.database.driver-class-name=org.postgresql.Driver
 # Reporting Directory and File Name
-#mandatory.reporting.directory=/MandatoryReporting
 mandatory.reporting.directory=src/data/MandatoryReporting
 mandatory.reporting.file=ReportingExample.csv
 # Covid Directory and File Ext
